chore: missing eslint fixes for modals
